+++ views/pages/admin/articles.php
@@ -2,15 +2,20 @@
 
 use frame\lists\paged\PagerModel;
 use engine\articles\Article;
-use engine\articles\ArticlePagedList;
 use engine\users\User;
 use function lightlib\shorten;
 use frame\tools\JsonEncoder;
 use frame\auth\InitAccess;
+use frame\config\ConfigRouter;
+use frame\database\Records;
+use frame\lists\base\IdentityList;
 
 InitAccess::accessRight('articles', 'see-list');
 $pagenumber = PagerModel::getRoutePage();
-$articles = new ArticlePagedList($pagenumber);
+$articleCount = Records::from(Article::getTable())->count('id');
+$articleLimit = ConfigRouter::getDriver()->findConfig('articles')->get('list.amount');
+$pager = new PagerModel($pagenumber, $articleCount, $articleLimit);
+$articles = new IdentityList(Article::class, ['id' => 'DESC'], $pager->getOffset(), $pager->getLimit());
 $rights = User::getMyRights('articles');
 
 $tableProps = ['items' => []];
@@ -31,13 +36,13 @@ $tableProps = JsonEncoder::forHtmlAttribute($tableProps);
 <div class="content__header">
     <div class="breadcrumbs">
         <span class="breadcrumbs__item breadcrumbs__item--current">
-            Статьи (<?= $articles->countOnPage() ?>)
+            Статьи (<?= $articles->count() ?>)
         </span>
     </div>
     <div class="actions">
-        <?php if ($articles->getPager()->countPages() > 1) : ?>
+        <?php if ($pager->countPages() > 1) : ?>
             <div class="actions__item">
-                <?php $articles->getPager()->show('admin') ?>
+                <?php $pager->show('admin') ?>
             </div>
         <?php endif ?>
         <?php if ($rights->can('add')): ?>